refactor: 인프라 관련 서비스 인터페이스 위치 변경
`web.global.thirdParty.aws.service.aws` -> `web.domain.application.service` 로 이동
